fifth

+++ src/main/java/com/validate/controller/CompanyController.java
@@ -19,7 +19,10 @@ public class CompanyController {
     @Autowired
     EmployeeRepo employeeRepo;
     
-    //gettting list of company
+    /**
+     *  getting list of company
+     * @return List<Company>
+     */
     @GetMapping
     public List<Company> get(){
         return companyRepo.findAll();
@@ -59,7 +62,10 @@ public class CompanyController {
     //getting companies by id
     @GetMapping("/{id}")
     public Company getById(@PathVariable Long id ) {
-    	return companyRepo.findById(id).get();
+    	 Optional<Company> findById = companyRepo.findById(id);
+    	 if(findById.isPresent())
+    		 return findById.get();
+    	 return null;
     }
     //getting list of employee by company id
     @GetMapping("/{id}/employees")
@@ -67,35 +73,86 @@ public class CompanyController {
     	return companyRepo.findById(id).get().getEmployees();
     	
     }
+    
     //adding employees in a company
     @PostMapping("/{id}/employees")
     public Employee addEmployee(@PathVariable("id")Long id,@RequestBody Employee employee) {
     	Company company =companyRepo.findById(id).get();
     	employee.setCompany(company);
-    	company.getEmployees().add(employee);
-    	companyRepo.save(company);
-    	return employee;
+//    	company.getEmployees().add(employee);
+//    	companyRepo.save(company);
+    	return employeeRepo.save(employee);
     }
 	
     //getting employee by id in a company
+//	  @GetMapping("/{id}/employees/{eid}") 
+//	  public Employee getByid(@PathVariable("id")Long id,@PathVariable("eid")Long eid) {
+//		  Optional <Company> com = companyRepo.findById(id);
+//		  Company byId= com.get();
+//		  Optional<Employee> emp=byId.getEmployees().stream().filter(e->e.getId().equals(eid)).findAny();
+//		  System.out.println(emp.isPresent());
+//		  
+//		  return emp.get();
+//	  }
 	  @GetMapping("/{id}/employees/{eid}") 
 	  public Employee getByid(@PathVariable("id")Long id,@PathVariable("eid")Long eid) {
-		  Optional <Company> com = companyRepo.findById(id);
-		  Company byId= com.get();
-		  Optional<Employee> emp=byId.getEmployees().stream().filter(e->e.getId().equals(eid)).findAny();
-		  System.out.println(emp.isPresent());
-		  
-		  return emp.get();
+		  return employeeRepo.findByIdAndCompanyId(eid, id);
 	  }
+	
+//	  //deleting employee 
+//	  @DeleteMapping("/{id}/employees/{eid}")
+//	  public String deleteByid(@PathVariable("id")Long id,@PathVariable("eid")Long eid) {
+//		 if(companyRepo.findById(id).isPresent()) {
+//			 if(employeeRepo.findById(eid).isPresent()) {
+//				 Employee employee = employeeRepo.findById(eid).get();
+//				 Company company = companyRepo.findById(id).get();
+//				 if(company.getEmployees().contains(employee)) {
+//					 company.getEmployees().remove(employee);
+//					 employeeRepo.delete(employee);
+//					 return "deleted successfully";
+//				 }else {
+//					 return "Not Authorized ";
+//				 }
+//			 }else {
+//				 return "Employee Doesn't exists";
+//			 }
+//		 }else {
+//			 return "company doesn't exist";
+//		 }
+//	  }
+//	  
 	  @DeleteMapping("/{id}/employees/{eid}")
-	  public void deleteByid(@PathVariable("id")Long id,@PathVariable("eid")Long eid) {
-		   employeeRepo.deleteById(eid);
-		   
+	  public String deleteByid(@PathVariable("id")Long id,@PathVariable("eid")Long eid) {
+		  Employee emp = employeeRepo.findByIdAndCompanyId(eid, id);
+		  if( emp!=null) {
+			  employeeRepo.deleteById(eid);
+			  return "deleted";
+		  }
+		  return "could not delete";
 	  }
+	  
+	  
+		  
+//	  
+//	  //editing employee
+//	  @PutMapping("/{id}/employees")
+//	  public Employee update(@PathVariable("id")Long id,@RequestBody Employee emp) {
+//		  if(companyRepo.findById(id).get().getEmployees().contains(emp)) {
+//		  emp.setCompany(companyRepo.findById(id).get());
+//		  return employeeRepo.save(emp);
+//		  }else {
+//			  return null;
+//		  }
+//
+//	  }
 	  @PutMapping("/{id}/employees")
-	  public Employee Update(@PathVariable("id")Long id,@RequestBody Employee emp) {
-		  emp.setCompany(companyRepo.findById(id).get());
-		  return employeeRepo.save(emp);
+	  public Employee updateEmp(@PathVariable("id")Long id,@RequestBody Employee emp) {
+		  Employee findByIdAndCompanyId = employeeRepo.findByIdAndCompanyId(emp.getId(), id);
+		  if(findByIdAndCompanyId!=null){
+			  emp.setCompany(companyRepo.findById(id).get());
+			  return employeeRepo.save(emp);
+		  }
+		  return null;
 	  }
 	
 }
